Refactorisation de la structure des fichiers du projet
