+++ Code/srcs/execution/execution_env_path.c
@@ -1,0 +1,60 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   execution_env_path.c                               :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: cmegret <cmegret@student.42lausanne.ch>    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/09/07 19:02:56 by cmegret           #+#    #+#             */
+/*   Updated: 2024/09/07 19:07:31 by cmegret          ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "../../header/Minishell.h"
+
+char	**get_env_paths(char **envp)
+{
+	int		i;
+	char	**paths;
+
+	i = 0;
+	while (envp[i])
+	{
+		if (ft_strncmp(envp[i], "PATH=", 5) == 0)
+		{
+			paths = ft_split(envp[i] + 5, ':');
+			return (paths);
+		}
+		i++;
+	}
+	return (NULL);
+}
+
+char	*get_cmd_path(char *cmd, char **envp)
+{
+	char	**paths;
+	char	*full_path;
+	int		i;
+
+	if (cmd == NULL || envp == NULL)
+		error_and_exit("cmd or envp is NULL");
+	paths = get_env_paths(envp);
+	if (!paths)
+		error_and_exit("get_env_paths failed");
+	i = 0;
+	while (paths[i])
+	{
+		full_path = ft_strjoin(ft_strjoin(paths[i], "/"), cmd);
+		if (!full_path)
+			error_and_exit("ft_strjoin failed");
+		if (access(full_path, X_OK) == 0)
+		{
+			free(paths);
+			return (full_path);
+		}
+		free(full_path);
+		i++;
+	}
+	free(paths);
+	return (NULL);
+}
