+++ Code/srcs/main/main_utils.c
@@ -1,29 +1,29 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   main.c                                             :+:      :+:    :+:   */
+/*   main_utils.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: cmegret <cmegret@student.42lausanne.ch>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/09/02 14:30:52 by cmegret           #+#    #+#             */
-/*   Updated: 2024/09/04 20:41:28 by cmegret          ###   ########.fr       */
+/*   Created: 2024/09/02 14:36:39 by cmegret           #+#    #+#             */
+/*   Updated: 2024/09/07 19:13:01 by cmegret          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../header/Minishell.h"
+#include "../../header/Minishell.h"
 
-int	main(int argc, char **argv, char **envp)
+void	ft_check_args(int argc, char **argv)
 {
-	char	*prompt;
-
-	ft_check_args(argc, argv);
-	prompt = readline("Minishell : ");
-	while (prompt != NULL)
+	if (argc > 1)
 	{
-		add_history(prompt);
-		execute_cmd(ft_parse(ft_tokenize(prompt)), envp);
-		free(prompt);
-		prompt = readline("Minishell : ");
+		fprintf(stderr, "Usage: %s\n", argv[0]);
+		fprintf(stderr, "No arguments are allowed\n");
+		exit(EXIT_FAILURE);
 	}
-	return (0);
+}
+
+void	error_and_exit(const char *message)
+{
+	perror(message);
+	exit(EXIT_FAILURE);
 }
