+++ Code/srcs/token/token_main.c
@@ -1,52 +1,44 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   tokenization.c                                     :+:      :+:    :+:   */
+/*   token_main.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: cmegret <cmegret@student.42lausanne.ch>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/09/02 14:47:54 by cmegret           #+#    #+#             */
-/*   Updated: 2024/09/04 09:48:15 by cmegret          ###   ########.fr       */
+/*   Created: 2024/09/06 17:19:58 by cmegret           #+#    #+#             */
+/*   Updated: 2024/09/07 19:20:00 by cmegret          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../header/Minishell.h"
+#include "../../header/Minishell.h"
 
-static int	handle_operator_token(char *prompt, int i, struct s_token **tok)
+static struct s_token	*ft_init_token(void)
+{
+	struct s_token	*tok;
+
+	tok = malloc(sizeof(struct s_token));
+	if (!tok)
+		return (NULL);
+	tok->next = NULL;
+	tok->str = NULL;
+	tok->type = 0;
+	tok->is_operator = NULL;
+	tok->is_delimited = false;
+	return (tok);
+}
+
+static int	ft_process_prompt(char *prompt, int i, struct s_token **tok)
 {
 	if (ft_previous_char_is_undelimited_operator(*tok))
 		return (ft_continue_operator_token(prompt, i, tok));
 	else if (ft_is_operator_character(prompt[i]))
 		return (ft_new_operator_token(prompt, i, tok));
-	return (0);
-}
-
-static int	handle_blank_token(char *prompt, int i, struct s_token **tok)
-{
-	if (ft_is_blank(prompt[i]))
+	else if (ft_is_blank(prompt[i]))
 		return (ft_tokenize_blank(tok));
-	return (0);
-}
-
-static int	handle_word_token(char *prompt, int i, struct s_token **tok)
-{
-	if (ft_previous_char_part_of_word(*tok))
+	else if (ft_previous_char_part_of_word(*tok))
 		return (ft_append_char_to_word(tok, prompt[i]));
 	else
 		return (ft_new_word(tok, prompt[i]));
-}
-
-static int	handle_token(char *prompt, int i, struct s_token **tok)
-{
-	int	increment;
-
-	increment = handle_operator_token(prompt, i, tok);
-	if (increment)
-		return (increment);
-	increment = handle_blank_token(prompt, i, tok);
-	if (increment)
-		return (increment);
-	return (handle_word_token(prompt, i, tok));
 }
 
 struct s_token	*ft_tokenize(char *prompt)
@@ -55,15 +47,14 @@ struct s_token	*ft_tokenize(char *prompt)
 	struct s_token	*head;
 	int				i;
 
-	tok = initialize_token();
+	tok = ft_init_token();
 	if (!tok)
 		return (NULL);
 	head = tok;
 	i = 0;
 	while (prompt[i])
-		i += handle_token(prompt, i, &tok);
+		i += ft_process_prompt(prompt, i, &tok);
 	if (!prompt[i])
 		tok = ft_tokenize_end_of_input(tok);
-	ft_print_all_token_strings(&head);
 	return (head);
 }
