+++ Code/srcs/token/token_operator.c
@@ -1,36 +1,20 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   tokenization_operator.c                            :+:      :+:    :+:   */
+/*   token_operator.c                                   :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: cmegret <cmegret@student.42lausanne.ch>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/09/04 09:23:27 by cmegret           #+#    #+#             */
-/*   Updated: 2024/09/04 09:24:17 by cmegret          ###   ########.fr       */
+/*   Created: 2024/09/06 17:27:19 by cmegret           #+#    #+#             */
+/*   Updated: 2024/09/06 17:32:20 by cmegret          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../header/Minishell.h"
+#include "../../header/Minishell.h"
 
-int	ft_is_operator_character(char c)
+static int	ft_is_double_greater(char *prompt, int i, struct s_token **tok)
 {
-	return (c == '>' || c == '<' || c == '|');
-}
-
-int	ft_previous_char_is_undelimited_operator(struct s_token *tok)
-{
-	if (tok->str && tok->is_operator == true && tok->is_delimited == false)
-		return (1);
-	else
-		return (0);
-}
-
-int	ft_continue_operator_token(char *prompt, int i, struct s_token **tok)
-{
-	if (ft_strlen((*tok)->str) != 1)
-		(*tok)->is_delimited = true;
-	else if (prompt[i] == '>' && prompt[i - 1] == '>'
-		&& (*tok)->is_quoted == false)
+	if (prompt[i] == '>' && prompt[i - 1] == '>' && (*tok)->is_quoted == false)
 	{
 		free((*tok)->str);
 		(*tok)->str = ft_strdup(">>");
@@ -38,8 +22,12 @@ int	ft_continue_operator_token(char *prompt, int i, struct s_token **tok)
 		(*tok)->is_delimited = true;
 		return (1);
 	}
-	else if (prompt[i] == '<' && prompt[i - 1] == '<'
-		&& (*tok)->is_quoted == false)
+	return (0);
+}
+
+static int	ft_is_double_less(char *prompt, int i, struct s_token **tok)
+{
+	if (prompt[i] == '<' && prompt[i - 1] == '<' && (*tok)->is_quoted == false)
 	{
 		free((*tok)->str);
 		(*tok)->str = ft_strdup("<<");
@@ -47,22 +35,38 @@ int	ft_continue_operator_token(char *prompt, int i, struct s_token **tok)
 		(*tok)->is_delimited = true;
 		return (1);
 	}
+	return (0);
+}
+
+int	ft_continue_operator_token(char *prompt, int i, struct s_token **tok)
+{
+	if (ft_strlen((*tok)->str) != 1)
+		(*tok)->is_delimited = true;
+	else if (ft_is_double_greater(prompt, i, tok))
+		return (1);
+	else if (ft_is_double_less(prompt, i, tok))
+		return (1);
 	else
 		(*tok)->is_delimited = true;
 	return (0);
 }
 
-int	ft_new_operator_token(char *prompt, int i, struct s_token **tok)
+static void	ft_set_operator_token(char c, struct s_token **tok)
 {
-	(*tok) = ft_create_new_token(*tok);
-	if (prompt[i] == '>')
+	if (c == '>')
 		(*tok)->str = ft_strdup(">");
-	else if (prompt[i] == '<')
+	else if (c == '<')
 		(*tok)->str = ft_strdup("<");
-	else if (prompt[i] == '|')
+	else if (c == '|')
 		(*tok)->str = ft_strdup("|");
 	(*tok)->is_operator = true;
 	(*tok)->is_delimited = false;
 	(*tok)->next = NULL;
+}
+
+int	ft_new_operator_token(char *prompt, int i, struct s_token **tok)
+{
+	(*tok) = ft_create_new_token(*tok);
+	ft_set_operator_token(prompt[i], tok);
 	return (1);
 }
